added component the sortline for sort, added alerts for delete and add a product...

+++ src/components/ProductsList/Products/Products.js
@@ -10,6 +10,7 @@ import {
   selectIsRefresh,
   getIsRefresh,
 } from "./Products-slice";
+import { SortLine } from "../SortLine/SortLine";
 import Button from "react-bootstrap/Button";
 import Modal from "react-bootstrap/Modal";
 import s from "./Products.module.css";
@@ -47,15 +48,29 @@ export function Products() {
   const handleShow = () => setShowModal(true);
 
   const deleteProduct = (e) => {
-    fetch(`http://localhost:8000/product/${e.target.id}`, {
-      method: "DELETE",
-    }).then((e) => e.json());
+    const attention = window.confirm(
+      "Are you sure you want to delete this product from the database?"
+    );
+
+    if (attention) {
+      fetch(`http://localhost:8000/product/${e.target.id}`, {
+        method: "DELETE",
+      }).then((e) => e.json());
+      // Save it!
+      console.log("Product was deleted from the database.");
+    } else {
+      // Do nothing!
+      console.log("Product was not deleted from the database.");
+    }
+
     dispatch(getIsRefresh(true));
   };
 
   return (
     <div className="container">
       <h1 className={s.title}>PRODUCT LIST:</h1>
+      <SortLine />
+
       <ul className="card-set list">
         {products.map((e) => (
           <li className="card-set__item" key={e.id} id={e.id}>
@@ -64,11 +79,18 @@ export function Products() {
               style={{ paddingTop: "20px", marginTop: "0px" }}
             >
               <li className="card-set__text">
-                <span style={{ fontWeight: "bold" }}>imageUrl: </span>
-                {e.imageUrl}
+                <img
+                  src={e.imageUrl}
+                  alt="Paris"
+                  width="320px"
+                  height="240px"
+                ></img>
+                {/* <span style={{ fontWeight: "bold" }}>imageUrl: </span> */}
+                {/* {e.imageUrl} */}
               </li>
               <li className="card-set__text">
-                <span style={{ fontWeight: "bold" }}>name: </span> {e.name}
+                {/* <span style={{ fontWeight: "bold" }}>name: </span> */}
+                <h3>{e.name}</h3>
               </li>
               <li className="card-set__text">
                 <span style={{ fontWeight: "bold" }}>count: </span>
